Create Challenge_Day4.py
This challenge was actually challenging.
It makes you think, and I like that. But as usual, I made it my b**ch :).
As you can see rom my first to revised code, most of the work happens in amIOld.
 __init__ is just for declarations. Also, remember the fullstops after the print statement, I kept getting a wrong answer output and after numerous and I mean numerous attempts, it hit me, "Check the Print Statements" and alas, it hit me that a fullstop was my problem. Sounds a bit philosophical though, "a fullstop was all I needed to end this" That line can be applied everywhere, code, relationships, life... think about it....

To conclude, it's easy, if I can do it, so can you.