Create Challenge_Day6.py
This one made me think. Which is good otherwise it's not a challenge now is it?
My output was always correct but coming out iffy. 
After numerous if,elifs and elses, it dawned on me,
I should be looking at this as an array. and abuse it's properties. mwahahahah!!
I have left two print options for the output that work.
The first one uses two print functions with a trailing comma, the comma allows one to print multiple statements on one line.
The second one, which is more common, is ye 'ol concatenation :)

Like I said, easy.