Create Challenge_Day7.py
You might think that this challenge needs a function but it's testing on you knowledge of arrays and how to manipulate them.
So, all you have to do is to print the output, the trick is how you do it. As you can see, I have used join() and map().
The reverse bit is simple, just put a negative(-) on your skip it runs in reverse. Enjoy :)